requirements.txt
pandas
sentence-transformers
faiss-cpu
transformers
torch
sentencepiece
urllib3<2
